Cast to correct type.
git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/sctplib1@742 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ sctplib/programs/chargen_server.c
@@ -1,5 +1,5 @@
 /*
- *  $Id: chargen_server.c,v 1.14 2005/08/03 11:24:31 dreibh Exp $
+ *  $Id$
  *
  * SCTP implementation according to RFC 2960.
  * Copyright (C) 2000 by Siemens AG, Munich, Germany.
@@ -448,9 +448,9 @@ int main(int argc, char **argv)
                                          noOfLocalAddresses, localAddressList,
                                          chargenUlp);
 
-    SCTP_getAssocDefaults(sctpInstance, &instanceParameters);
+    SCTP_getAssocDefaults((unsigned short)sctpInstance, &instanceParameters);
     instanceParameters.maxSendQueue = SEND_QUEUE_SIZE;
-    SCTP_setAssocDefaults(sctpInstance, &instanceParameters);
+    SCTP_setAssocDefaults((unsigned short)sctpInstance, &instanceParameters);
 
     /* run the event handler forever */
     while (1) {
