+++ sctplib/programs/discard_server.c
@@ -1,5 +1,5 @@
 /*
- *  $Id: discard_server.c,v 1.9 2005/03/08 13:29:52 tuexen Exp $
+ *  $Id$
  *
  * SCTP implementation according to RFC 2960.
  * Copyright (C) 2000 by Siemens AG, Munich, Germany.
@@ -433,10 +433,10 @@ int main(int argc, char **argv)
                                          noOfLocalAddresses, localAddressList,
                                          discardUlp);
     
-    SCTP_getAssocDefaults(sctpInstance, &instanceParameters);
+    SCTP_getAssocDefaults((unsigned short)sctpInstance, &instanceParameters);
     if (myRwndSpecified)
         instanceParameters.myRwnd = myRwnd;
-    SCTP_setAssocDefaults(sctpInstance, &instanceParameters);
+    SCTP_setAssocDefaults((unsigned short)sctpInstance, &instanceParameters);
 
     
     /* run the event handler forever */
